Fix test fixtures.
